// src/utils/types.ts
export type APIStrut = {
    velocity: number;
    altitude: number;
    temperature: number;
    statusMessage: string;
    isAscending: boolean;
    isActionRequired: boolean;
};

export type APIStrutWS = {
    Velocity: number;
    Altitude: number;
    Temperature: number;
    StatusMessage: string;
    IsAscending: boolean;
    IsActionRequired: boolean;
};